# Conflicts:
#	Agrotutor/Agrotutor.iOS/Agrotutor.iOS.csproj

+++ Agrotutor/Agrotutor/Modules/Map/ViewModels/MapPageViewModel.cs
@@ -2,6 +2,8 @@ using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
+using System.Net.Http;
+using System.Text;
 using System.Threading.Tasks;
 using Acr.UserDialogs;
 using Agrotutor.Core;
@@ -24,7 +26,6 @@ using Agrotutor.Modules.Plot.ViewModels;
 using Agrotutor.Modules.PriceForecasting.Types;
 using Agrotutor.Modules.Weather;
 using Agrotutor.Modules.Weather.Types;
-using Castle.Core.Internal;
 using Microsoft.Extensions.Localization;
 using Plugin.Permissions;
 using Plugin.Permissions.Abstractions;
@@ -44,6 +45,7 @@ using Agrotutor.Modules.Plot.Views;
 using Agrotutor.Modules.PriceForecasting.ViewModels;
 using Agrotutor.Modules.Weather.ViewModels;
 using Agrotutor.Modules.Weather.Views;
+using Flurl.Http;
 
 namespace Agrotutor.Modules.Map.ViewModels
 {
@@ -152,6 +154,10 @@ namespace Agrotutor.Modules.Map.ViewModels
         private bool showDeletePlotButton;
         private double selectedPlotActivityCost;
 
+        private readonly double tileSize = 130;
+        private string tileName = "Guanajuato";
+        private string _lastUploadDateString;
+
         public MapPageViewModel(
             INavigationService navigationService,
             IAppDataService appDataService,
@@ -176,7 +182,10 @@ namespace Agrotutor.Modules.Map.ViewModels
 
             Polygons = new ObservableCollection<Polygon>();
             Pins = new ObservableCollection<Pin>();
-            DownloadDeleteCommand = new DelegateCommand(DownloadTiles);
+            DownloadDeleteCommand=new DelegateCommand(async () =>
+            {
+                await DownloadTilesAsync();
+            });
             CheckDownloadStatus();
         }
 
@@ -211,26 +220,94 @@ namespace Agrotutor.Modules.Map.ViewModels
                 IsOfflineBasemapLayerEnabled = false;
             }
         }
-        public void DownloadTiles()
+
+        private async Task UpLoadDataAsync()
         {
+            try
+            {
+                TimeSpan dateRes;
+                if (_lastUploadDateString != null)
+                {
+                    DateTime uploadDateTime=DateTime.Parse(_lastUploadDateString);
+                     dateRes = DateTime.UtcNow.Subtract(uploadDateTime);
+                }
+
+                if (_lastUploadDateString == null||dateRes.Days >= 90)
+                {
+                    if (Connectivity.NetworkAccess == NetworkAccess.Internet)
+                    {
+                        HttpClient client = new HttpClient();
+                        if (Plots.ToList().Count > 0)
+                        {
+                            foreach (var plot in Plots)
+                            {
+                                var jsonObj = JsonConvert.SerializeObject(plot);
+                                var res = await client.PostAsync(Constants.UploadDataUrl,
+                                    new StringContent(jsonObj, Encoding.UTF8, "application/json"));
+                                if (!res.IsSuccessStatusCode)
+                                {
+                                    //todo:show notification or log the result
+                                    return;
+                                }
+
+                            }
+
+                            LastUploadDateString = DateTime.UtcNow.ToString();
+                        }
+                    }
+                }
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+        
+            }
+
+
+        }
+
+        public async Task DownloadTilesAsync( )
+        {
+      
             if (FileManager.CacheFileExists(Constants.DownloadTileUrl))
             {
-                // remove files
-                FileManager.DeleteOfflineCache(Constants.DownloadTileUrl);
-                //  LayerService.UpdateIsChecked( FileManager.GetCacheFilePath(downloadUrl));
-                DownloadStatusImage = "ic_download";
-                IsOfflineBasemapLayerEnabled = false;
-                OfflineBasemapLayerVisible = false;
+                var deleteOfflineMapMessage =
+                    string.Format(StringLocalizer.GetString("delete_offline_map_message"), tileSize);
+                var confirm = await MaterialDialog.Instance.ConfirmAsync(deleteOfflineMapMessage, null, StringLocalizer.GetString("download_offline_map_yes"),
+                    StringLocalizer.GetString("download_offline_map_cancel"));
+                if (confirm.Value)
+                {
+                    // remove files
+                    FileManager.DeleteOfflineCache(Constants.DownloadTileUrl);
+                    //  LayerService.UpdateIsChecked( FileManager.GetCacheFilePath(downloadUrl));
+                    DownloadStatusImage = "ic_download";
+                    IsOfflineBasemapLayerEnabled = false;
+                    OfflineBasemapLayerVisible = false;
+                }
+   
             }
             else
             {
-                TaskDownloadTiles();
+             
+                var downloadOfflineMapMessage =
+                    string.Format(StringLocalizer.GetString("download_offline_map_message"), tileSize);
+                var confirm = await MaterialDialog.Instance.ConfirmAsync(downloadOfflineMapMessage, null, StringLocalizer.GetString("download_offline_map_yes"),
+                    StringLocalizer.GetString("download_offline_map_cancel"));
+                if (confirm.Value)
+                {
+             
+               await TaskDownloadTiles();
+                    
+                }
+            
             }
         }
-        private void TaskDownloadTiles()
+        private async Task TaskDownloadTiles()
         {
             if (Connectivity.NetworkAccess == NetworkAccess.Internet)
             {
+                var downloadOfflineMapInProgress = string.Format(StringLocalizer.GetString("download_offline_map_in_progress"), tileName);
+                var loadingDialog = await MaterialDialog.Instance.LoadingDialogAsync(downloadOfflineMapInProgress);
                 IsDownloadButtonEnabled = false;
                 IDownloadFile file = CrossDownloadManager.Current.CreateDownloadFile(Constants.DownloadTileUrl);
 
@@ -247,6 +324,7 @@ namespace Agrotutor.Modules.Map.ViewModels
                                 IsOfflineBasemapLayerEnabled = true;
                                 LayerService.UpdateIsChecked(FileManager.GetCacheFilePath(Constants.DownloadTileUrl));
                                 System.Console.WriteLine("Downloading finished. " + file.DestinationPathName);
+                                loadingDialog.DismissAsync();
                                 break;
 
                             case DownloadFileStatus.FAILED:
@@ -292,6 +370,15 @@ namespace Agrotutor.Modules.Map.ViewModels
             {
                 SetProperty(ref _showSatelliteLayer, value);
                 Preferences.Set(Constants.ShowSatelliteTileLayerVisiblePreference, value);
+            }
+        }
+        public string LastUploadDateString
+        {
+            get => _lastUploadDateString;
+            set
+            {
+                SetProperty(ref _lastUploadDateString, value);
+                Preferences.Set(Constants.LastUploadDatePreference, value);
             }
         }
         public ObservableCollection<Polygon> Polygons
@@ -350,7 +437,7 @@ namespace Agrotutor.Modules.Map.ViewModels
         public DelegateCommand DownloadDeleteCommand { get; set; }
         public DelegateCommand ShowCurrentPlotCost => new DelegateCommand(async () =>
         {
-            if (SelectedPlot?.BemData == null || SelectedPlot.BemData.Cost.IsNullOrEmpty())
+            if (SelectedPlot?.BemData == null || SelectedPlot.BemData.Cost==null||SelectedPlot.BemData.Cost.Count==0)
             {
                 await MaterialDialog.Instance.SnackbarAsync(StringLocalizer.GetString("cost_data_not_available"));
                 return;
@@ -364,7 +451,7 @@ namespace Agrotutor.Modules.Map.ViewModels
 
         public DelegateCommand ShowCurrentPlotIncome => new DelegateCommand(async () =>
         {
-            if (SelectedPlot?.BemData == null || SelectedPlot.BemData.Income.IsNullOrEmpty())
+            if (SelectedPlot?.BemData == null || SelectedPlot.BemData.Income==null||SelectedPlot.BemData.Income.Count==0)
             {
                 await MaterialDialog.Instance.SnackbarAsync(StringLocalizer.GetString("income_data_not_available"));
                 return;
@@ -378,7 +465,7 @@ namespace Agrotutor.Modules.Map.ViewModels
 
         public DelegateCommand ShowCurrentPlotProfit => new DelegateCommand(async () =>
         {
-            if (SelectedPlot?.BemData == null || SelectedPlot.BemData.Profit.IsNullOrEmpty())
+            if (SelectedPlot?.BemData == null || SelectedPlot.BemData.Profit==null|| SelectedPlot.BemData.Profit.Count==0)
             {
                 await MaterialDialog.Instance.SnackbarAsync(StringLocalizer.GetString("profit_data_not_available"));
                 return;
@@ -392,7 +479,7 @@ namespace Agrotutor.Modules.Map.ViewModels
 
         public DelegateCommand ShowCurrentPlotYield => new DelegateCommand(async () =>
         {
-            if (SelectedPlot?.BemData == null || SelectedPlot.BemData.Yield.IsNullOrEmpty())
+            if (SelectedPlot?.BemData == null || SelectedPlot.BemData.Yield==null||SelectedPlot.BemData.Yield.Count==0)
             {
                 await MaterialDialog.Instance.SnackbarAsync(StringLocalizer.GetString("yield_data_not_available"));
                 return;
@@ -686,10 +773,10 @@ namespace Agrotutor.Modules.Map.ViewModels
 
                         case MapTask.DelineationNotEnoughPoints:
                         case MapTask.DelineationEnoughPoints:
-                            if (CurrentDelineation == null) CurrentDelineation = new List<Position>();
+                            if (CurrentDelineation == null) CurrentDelineation = new List<DelineationPosition>();
 
                             var pos = Position.From(args.Point);
-                            CurrentDelineation.Add(pos);
+                            CurrentDelineation.Add(new DelineationPosition{Position= pos });
                             var pin = new Pin
                             {
                                 Position = Position.From(args.Point).ForMap(),
@@ -927,7 +1014,7 @@ namespace Agrotutor.Modules.Map.ViewModels
             }
         }
 
-        public List<Position> CurrentDelineation { get; set; }
+        public List<DelineationPosition> CurrentDelineation { get; set; }
 
         public HubFeature CurrentHubContact
         {
@@ -1365,7 +1452,7 @@ namespace Agrotutor.Modules.Map.ViewModels
                 {
                     SelectedPlot.Delineation = CurrentDelineation;
                     AppDataService.UpdatePlotAsync(SelectedPlot);
-                    CurrentDelineation = new List<Position>();
+                    CurrentDelineation = new List<DelineationPosition>();
                     CurrentMapTask = MapTask.Default;
                     EndDelineation();
                 });
@@ -1374,7 +1461,7 @@ namespace Agrotutor.Modules.Map.ViewModels
             new DelegateCommand(
                 () =>
                 {
-                    CurrentDelineation = new List<Position>();
+                    CurrentDelineation = new List<DelineationPosition>();
                     CurrentMapTask = MapTask.Default;
                     EndDelineation();
                 });
@@ -1469,7 +1556,8 @@ namespace Agrotutor.Modules.Map.ViewModels
                         {
                             Tag = plot.Position
                         };
-                        foreach (var position in positions) polygon.Positions.Add(position.ForMap());
+                        foreach (var position in positions)
+                            polygon.Positions.Add(position.Position.ForMap());
                         Polygons.Add(polygon);
                     }
                 }
@@ -1481,7 +1569,7 @@ namespace Agrotutor.Modules.Map.ViewModels
             if (CurrentDelineation.Count < 3) return;
             RemoveDelineationPolygon();
             var polygon = new Polygon { FillColor = Color.Transparent };
-            foreach (var position in CurrentDelineation) polygon.Positions.Add(position.ForMap());
+            foreach (var position in CurrentDelineation) polygon.Positions.Add(position.Position.ForMap());
             CurrentPolygon = polygon;
             Polygons.Add(polygon);
         }
@@ -1657,7 +1745,7 @@ namespace Agrotutor.Modules.Map.ViewModels
 
                 //tasks.Add(LoadPlots());
                 await LoadPlots();
-
+                await UpLoadDataAsync();
                 //tasks.Add(RefreshWeatherData());
                 await RefreshWeatherData();
 
@@ -1908,6 +1996,7 @@ namespace Agrotutor.Modules.Map.ViewModels
                 Preferences.Get(Constants.InvestigationPlatformsLayerVisiblePreference, false);
                 OfflineBasemapLayerVisible = Preferences.Get(Constants.OfflineBasemapLayerVisiblePreference, false);
                 ShowSatelliteTileLayer = Preferences.Get(Constants.ShowSatelliteTileLayerVisiblePreference, true);
+                LastUploadDateString = Preferences.Get(Constants.LastUploadDatePreference,null);
             }
         }
